src/requirements.txt
pandas
boto3
xlrd
psycopg2-binary